Replace NewRunner and remove read config (move to main)
