+++ network-discovery/config/logger.go
@@ -1,8 +1,6 @@
 package config
 
 import (
-	"flag"
-	"fmt"
 	"log/slog"
 	"os"
 	"strings"
@@ -35,29 +33,4 @@ func NewLogger(logLevel string, logFormat string) *slog.Logger {
 	}
 
 	return slog.New(h)
-}
-
-// ReadConfigFile reads required config file
-func ReadConfigFile() ([]byte, error) {
-
-	configPath := flag.String("config", "", "path to the configuration file (required)")
-
-	flag.Parse()
-
-	if *configPath == "" {
-		fmt.Fprintf(os.Stderr, "Usage of network-discovery:\n")
-		flag.PrintDefaults()
-		return nil, fmt.Errorf("")
-
-	}
-	if _, err := os.Stat(*configPath); os.IsNotExist(err) {
-		return nil, fmt.Errorf("configuration file '%s' does not exist", *configPath)
-	}
-
-	fileData, err := os.ReadFile(*configPath)
-	if err != nil {
-		return nil, fmt.Errorf("error reading configuration file: %v", err)
-	}
-
-	return fileData, nil
 }
