+++ network-discovery/docker/Dockerfile
@@ -21,8 +21,8 @@ COPY --from=builder /src/network-discovery/build/network-discovery /usr/local/bi
 
 LABEL maintainer="techops@netboxlabs.com"
 LABEL org.opencontainers.image.title="Orb Network Discovery"
-LABEL org.opencontainers.image.url="https://github.com/netboxlabs/orb-discovery/network-discovery"
-LABEL org.opencontainers.image.documentation="https://github.com/netboxlabs/orb-discovery/network-discovery"
+LABEL org.opencontainers.image.url="https://github.com/netboxlabs/orb-discovery/tree/develop/network-discovery"
+LABEL org.opencontainers.image.documentation="https://github.com/netboxlabs/orb-discovery/tree/develop/network-discovery"
 LABEL org.opencontainers.image.description="NetBox Labs network-discovery image"
 LABEL org.opencontainers.image.vendor="NetBox Labs, Inc"
 LABEL org.opencontainers.image.authors="techops@netboxlabs.com"
