+++ network-discovery/version/version.go
@@ -5,22 +5,22 @@ import (
 	"strings"
 )
 
-// Version is the version of the diode-server
+// Version is the version of the network-discovery
 //
 //go:embed BUILD_VERSION.txt
 var buildVersion string
 
-// Commit is the commit of the diode-server
+// Commit is the commit of the network-discovery
 //
 //go:embed BUILD_COMMIT.txt
 var buildCommit string
 
-// GetBuildVersion returns the build version of the diode-server
+// GetBuildVersion returns the build version of the network-discovery
 func GetBuildVersion() string {
 	return strings.TrimSpace(buildVersion)
 }
 
-// GetBuildCommit returns the build commit of the diode-server
+// GetBuildCommit returns the build commit of the network-discovery
 func GetBuildCommit() string {
 	return strings.TrimSpace(buildCommit)
 }
